requirements/common.txt
django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
Pillow==11.0.0
qrcode==8.0